refactor: Refactor every parts to use serverless backend to exchange messages and calculate sentiment score
