+++ src/app/api/messages/route.js
@@ -1,0 +1,10 @@
+import { NextResponse } from 'next/server';
+
+export async function POST(req) {
+    return new NextResponse({
+        status: 200,
+        body: {
+            messages: []
+        }
+    })
+}
