+++ src/components/Layout.js
@@ -8,7 +8,7 @@ const Layout = (props) => {
             <meta charSet="utf-8" />
             <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
             <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossOrigin="anonymous" />
-            <title>{props.pageTitle || 'Realtime Chat'}</title>
+            <title>{props.pageTitle || 'Riffle Ripple'}</title>
         </Head>
         {props.children}
     </Fragment>
